[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,7 +2,7 @@
 
 int main()
 {
-	long long num1, num2;
+	int num1, num2;
 	std::cin >> num1 >> num2;
 
 	std::cout << (num1 + num2) * (num1 - num2);
